requirements.txt
langchain>=1.2.0
langchain-classic>=1.0.0
langchain-community>=0.4.0
langchain-groq>=0.1.0
langchain-text-splitters>=1.1.0
faiss-cpu>=1.13.0
sentence-transformers>=3.4.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.1.0
pydantic>=2.12.0
streamlit>=1.45.0
requests>=2.32.0
